Refactor block keyboard shortcuts to fix e2e tests
